Use automatic wiring for event deduplicator.
